llm_server/requirements.txt
google-generativeai
grpcio
grpcio-tools
protobuf
kafka-python
python-dotenv
